requirements.txt
# ============================================================
# PRISMA - Dependencies
# Python 3.12+
# ============================================================

# GUI Framework
flet>=0.25.0

# Web Scraping (Playwright + HTML to Markdown)
playwright>=1.49.0
html2text>=2024.2.26
httpx>=0.28.0

# AI Inference (Local LLM)
llama-cpp-python>=0.3.2

# File Processing
python-docx>=1.1.0
PyPDF2>=3.0.1

# Utilities
aiofiles>=24.1.0
rich>=13.9.0
pydantic>=2.10.0
huggingface-hub>=0.27.0

# Development (optional)
# pyinstaller>=6.11.0
